==> 1_data-cleaning.Rmd <==

```{r}
library(dplyr)
library(stringr)
library(purrr)

# load data 

filename = "/Users/kelsey/Dropbox (UMass Medical School)/Kelsey/Lassa Fever/Data/commcare-cleaned-data-2018-2021.csv"
lassa <- read.csv(filename)

age <- lassa$basic_information.age
sex <- lassa$basic_information.sex
pcrct <- lassa$pcr_list.pcr_cycle_threshold.0
dofbp <- lassa$clinical_features.before_admission_group.before_admission_symptom_duration.fever_duration.symptom_before_admission %>%
  str_count("D")	
temp <- lassa$clinical_features.at_admission_group.vital_signs.at_admission_temperature
sbp	<- lassa$clinical_features.at_admission_group.vital_signs.at_admission_systolic_blood_pressure
dbp	<- lassa$clinical_features.at_admission_group.vital_signs.at_admission_diastolic_blood_pressure
pr <- lassa$clinical_features.at_admission_group.vital_signs.at_admission_pulse_rate
rr <- lassa$clinical_features.at_admission_group.vital_signs.at_admission_respiratory_rate
temp_max <- lassa$clinical_factors.vital_signs.max_temperature.d1
pr_max <- lassa$clinical_factors.vital_signs.max_pulse_rate.d1
sbp_max <- lassa$clinical_factors.vital_signs.max_systolic_blood_pressure.d1
dbp_max <- lassa$clinical_factors.vital_signs.max_diastolic_blood_pressure.d1
rr_max <- lassa$clinical_factors.vital_signs.max_respiratory_rate.d1

####################################################################################################
# Group variables for severe cns and bleeding

dur <- lassa %>%
  dplyr::select(contains("_during_admission"))
dur <- map_df(dur, function(x) ifelse(grepl("d1", x), 1, 0))

bleeding_vars <- c("eyes_bleeding", "hematemesis_bleeding", "hematochezia_bleeding", "hemoptysis_bleeding", "injection_sites_bleeding", "mouth_bleeding", "nostrils_bleeding", "vaginal_bleeding", "other_bleeding")
scns_vars <- c("coma", "seizures", "tremors", "confusion")

bleeding <- dur %>%
  dplyr::select(contains(bleeding_vars))
bleeding <- data.frame(rowSums(bleeding))
bleeding[bleeding > 1] <- 1
blding <- bleeding$rowSums.bleeding.

# subset severe cns variables and sum values in new variable
scns <- dur %>%
  dplyr::select(contains(scns_vars))
scns <- data.frame(rowSums(scns))
scns[scns > 1] <- 1
scns <- scns$rowSums.scns.

# define other binary variables
out <- lassa$clinical_outcome.outcome
sthr <- dur$clinical_factors.during_admission_symptom_duration.sore_throat_duration.symptom_during_admission
coug <- dur$clinical_factors.during_admission_symptom_duration.cough_duration.symptom_during_admission
vomit <- dur$clinical_factors.during_admission_symptom_duration.vomiting_duration.symptom_during_admission
diarr <- dur$clinical_factors.during_admission_symptom_duration.diarrhea_duration.symptom_during_admission
head <- dur$clinical_factors.during_admission_symptom_duration.headache_duration.symptom_during_admission
abdp <- dur$clinical_factors.during_admission_symptom_duration.abdominal_pain_duration.symptom_during_admission
chesp <- dur$clinical_factors.during_admission_symptom_duration.chest_pain_duration.symptom_during_admission
weak <- dur$clinical_factors.during_admission_symptom_duration.weakness_duration.symptom_during_admission
jaun <- dur$clinical_factors.during_admission_symptom_duration.jaundice_duration.symptom_during_admission
reyes <- dur$clinical_factors.during_admission_symptom_duration.red_eyes_duration.symptom_during_admission
swell <- dur$clinical_factors.during_admission_symptom_duration.face_or_neck_swelling_duration.symptom_during_admission

################################################################################################
# only include blood chemistry tests administered within one day of admission
lassa1 <- lassa %>%
  dplyr::mutate(across(starts_with("blood"),
                      ~ifelse(blood_chemistry_list.days_since_admission_of_test.0 %in% 
                                c("0 days", "1 days"), ., NA))) %>%
  dplyr::mutate(across(starts_with("hematology"),
                      ~ifelse(hematology_list.days_since_admission_of_test.0 %in% 
                                c("0 days", "1 days"), ., NA))) %>%
  dplyr::mutate(across(starts_with("pcr"),
                      ~ifelse(pcr_list.days_since_admission_of_test.0 %in% 
                                c("0 days", "1 days"), ., NA)))

esr <- lassa1$hematology_list.erythrocyte_sedimentation_rate.0
k <- lassa1$blood_chemistry_list.potassium.0
na <- lassa1$blood_chemistry_list.sodium.0
alk <- lassa$blood_chemistry_list.alkaline_phosphatase.0
tpro <- lassa1$blood_chemistry_list.total_protein.0
alb <- lassa1$blood_chemistry_list.albumin.0
tbil <- lassa1$blood_chemistry_list.total_bilirubin.0
ast <- lassa1$blood_chemistry_list.aspartate_aminotransferase_ast.0
alt <- lassa1$blood_chemistry_list.alaninea_aminotransferas_alt.0
bun <- lassa1$blood_chemistry_list.blood_urea_nitrogen.0
cre <- lassa1$blood_chemistry_list.creatinine.0
wbc <- lassa1$hematology_list.white_cell_count.0
hct <- lassa1$hematology_list.haematocrit_concentration.0
plt <- lassa1$hematology_list.platelet.0
lym <- lassa1$hematology_list.lymphocytes.0
mon <- lassa1$hematology_list.monocytes.0
gra <- lassa1$hematology_list.granulocytes_neutrophils.0

########################################################################################################33
# reformat variables

sex <- case_when(sex == "male" ~ 0,
          sex == "female" ~ 1)

out <- case_when(out == "recovered" ~ 0,
          out == "died" ~ 1)

pcrct <- ifelse(pcrct > 60, NA, pcrct)


########## Add treatment variables
tx <- data.frame(matrix(NA, nrow = 841, ncol = 0))
tx$dialysis <- lassa$special_interventions.dialysis
tx$blood_trans <- lassa$special_interventions.blood_transfusion
tx$vent_support <- lassa$special_interventions.ventilatory_support.ventilatory_support_yes_no
tx$oxygen <- lassa$special_interventions.entilatory_support.oxygen
tx$intubation <- lassa$special_interventions.ventilatory_support.intubation_and_ventilation
tx$cardiac_support <- lassa$special_interventions.cardiac_support.cardiac_support_yes_no
tx$diuretics <- lassa$special_interventions.cardiac_support.diuretics
tx$ionotropes <- lassa$special_interventions.cardiac_support.ionotropes
tx$ribavirin <- lassa$special_interventions.treatment.ribavirin
tx$antibiotic <- lassa$antibiotics_group.antibiotics

tx[tx == ""] <- "no"

tx <- data.frame(lapply(tx, function(x) case_when(x == "no" ~ 0, x == "yes" ~ 1)))
#############################################################################################################
#### add antibiotic variables #####
type <- data.frame(ab = lassa$antibiotics_group.antibiotics_type)
atb <- c("Ceftriaxone", "Sulbactam", "Augmentin", "Levofloxacin", "Metronidazole",
         "Ampicillin", "Ciprofloxacine", "Cefuroxime", "Ceftazidime", "Azithromycin",
         "Doxycycline", "Amoxicillin")

ab <- bind_cols(type, map_dfc(atb, ~transmute(type, !!.x := grepl(.x, ab))))
ab <- data.frame(lapply(ab, function(x) case_when(x == "FALSE" ~ 0, x == "TRUE" ~ 1)))

#############################################################################################################
# save data
new <- data.frame(cbind(age, pcrct, dofbp, temp, sbp, dbp, pr, rr, temp_max, sbp_max, dbp_max, rr_max, esr, k,  na, alk, tpro, alb, tbil, ast, alt, bun, cre, wbc, hct, plt, lym, mon, gra, sex, scns, blding, sthr, coug, vomit, diarr, head, abdp, chesp, weak, jaun, reyes, swell, tx, ab, out))

lassa <- new %>% dplyr::mutate_all(na_if,"")

write.csv(lassa, "/Users/kelsey/Dropbox (UMass Medical School)/Kelsey/Lassa Fever/Data/2018-2021-cleaned.csv", row.names = F)
```
