==> 3_binary_variable_analysis.Rmd <==
---
title: "Descriptive analysis of binary variables in Lassa virus dataset"
---

```{r}
library(dplyr)
library(questionr)
```

Load data. 
```{r}
filename = "/Users/kelsey/Dropbox (UMass Medical School)/Kelsey/Lassa Fever/Data/2018-2021-cleaned.csv"
lassa <- read.csv(filename, header = TRUE, sep = ",")

bin <- lassa %>%
  dplyr::select(sex:swell)
bin$out <- as.factor(lassa$out)

died <- bin %>% filter(out == 1) %>% select(-out)
surv <- bin %>% filter(out == 0) %>% select(-out)
all <- rbind(died, surv)

list_df <- list(died, surv, all) 
```

Calculate totals and percents of people who experienced each symptom for all, died, and survived.
```{r}
total <- map_df(list_df, function(x) colSums(x, na.rm=T))
percent <- rbind(total[1,]/nrow(died)*100, total[2,]/nrow(surv)*100, total[3,]/nrow(all)*100)

died_string <- paste(total[1,],"/",nrow(died)," (",round(percent[1,],2),"%)", sep="")
surv_string <- paste(total[2,],"/",nrow(surv)," (",round(percent[2,],2),"%)", sep="")
all_string <- paste(total[3,],"/",nrow(all)," (",round(percent[3,],2),"%)", sep="")
```


Create separate logistic regression model for each variable and calculate odds ratios by exponentiating coefficients.
```{r}
glm <- map(bin[-19], function(x) glm(out ~ x, data = bin, family = binomial))
odds_df <- data.frame(map(glm, coefficients))
odds_vector <- exp(as.vector(odds_df[-1,]))
odds_rounded_list <- as.list(round(odds_vector,2))
#Calculate p-values for odds ratios
p_value_list <- map(map(glm, odds.ratio), "p")
p_value <- map(map(p_value_list, 2), function(x) round(x, 5))
```

Create a summary table of the data.

```{r}
binary_table <- data.frame(cbind(names(all), all_string, surv_string,  died_string, as.character(p_value), as.character(odds_rounded_list)))
names(binary_table) <- c("Symptom", "Overall", "Survived", "Died", "p_value", "odds_ratio") 
binary_table
# save output to csv
dir.create("lassa_fever_analysis")
write.csv(binary_table, "/Users/kelsey/Dropbox (UMass Medical School)/Kelsey/Lassa Fever/Data/lassa_fever_analysis/symptoms_at_admission.csv")
```

#```{r}
p <- round(data.frame(unlist(p_value)), 6)
newdata <- cbind(cfba_binary, OP =  all_percent, SP =  S_percent, DP =  D_percent, p = p)
names(newdata) <- c("name", "Overall", "Survived", "Died",  "p")
```

Order according to increasing p value and generate bar plot
#```{r}
newdata$name <- factor(newdata$name, levels = newdata$name[order(newdata$p)])
nd <- newdata %>%
  gather("Type", "Value", -name)
Type = c("Overall", "Died", "Survived", "p")
nd$Type <- factor(nd$Type, ordered = TRUE,  levels = Type)
```

#```{r}
p <- ggplot(subset(nd, Type %in% c("Overall" , "Died", "Survived")), aes(name, Value, fill = Type)) +
  geom_bar(position = "dodge", stat = "identity", colour="black") +
  scale_fill_manual(values=c("palegreen3", "coral2", "steelblue2")) +
  theme_bw() +
  ylab("Prevalence (%)") +
  xlab("Complication during treatment in hospital") +
  ylim(0, 100) +
  ggtitle("Prevalence of complications in patients diagnosed with Lassa fever") + 
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
# save output to pdf
pdf("./lassa_fever_analysis/symptoms_before_admission_prevalence.pdf")
p
dev.off()
```
