==> 4_continuous_variable_analysis.Rmd <==
---
title: "Descriptive analysis of continuous variables in Lassa virus data set"
---

```{r}
library(dplyr)
library(Rmisc)
library(questionr)
library(tidyr)
library(purrr)
```

Load data. 
```{r}
filename = "/Users/kelsey/Dropbox (UMass Medical School)/Kelsey/Lassa Fever/Data/2018-2021-cleaned.csv"
lassa <- read.csv(filename, header = TRUE, sep = ",")

cont <- lassa[c(1:29)]
cont <- cbind(cont, "out" = as.factor(lassa$out))

died <- cont %>% filter(out == 1) %>% select(-out)
surv <- cont %>% filter(out == 0) %>% select(-out)
all <- rbind(died,surv)

list_df <- list(died, surv, all)
```

Calculate mean and 95% CI for survived and died. Gives error message "NaNs produced" because some variables (calcium and phosphates) have no data. I left these in because in the real dataset these variables may have values.

```{r}
#write function to omit NA when calculating CI
CI_f <- function(x) {
  CI(na.omit(x), ci=.95)
}
#function to extract the mean, upper, and lower CI and save to a data frame
mean_CI <-function(x) {
  map_df(x, ~list(mean = CI_f(.x)["mean"], 
                      lower = CI_f(.x)["lower"],
                      upper = CI_f(.x)["upper"]))
}
#make lists of mean and CI for survived and died
means <- map(list_df, ~round(mean_CI(.x), 2))
names(means) <- c("died_mean", "survived_mean", "all_mean")
#write function to create string
string <- function(data) {
  paste0(data$mean, " (", data$lower, "-", data$upper, ")")
}
string_means <- map(means, string)
```


Calculate missing observations and percent missing for each variable.

```{r}
missing <- map_int(all, ~sum(is.na(.x)))
missing_percent <- round(missing/nrow(all) *100,  0)
missing_string <- paste(missing,"/",nrow(all), " (",missing_percent,"%)",sep = "")

data <- cont %>% filter(!is.na(out))
```

Divide age by 10 and other variables by IQR. Then calculate odds ratio. 

```{r}
data_scaled <-  cbind("out" = data$out,
                     "age" = data$age/10, 
                     data.frame(map(data[-c(1,30)], function(x) x/IQR(x,na.rm = TRUE))))
glm <- map(data_scaled[-1], function(x) glm(out ~ x, data = data_scaled, family = binomial))
```

```{r}
odds <- map(glm, odds.ratio)
odds_CI_2.5 <- round(map_dbl(map(odds, "2.5 %"), 2), 2)
odds_CI_97.5 <- round(map_dbl(map(odds, "97.5 %"), 2), 2)
OR <- round(map_dbl(map(odds, "OR"), 2), 2)
p_value <- format(signif(map_dbl(map(odds, "p"), 2), 2), scientific = FALSE)
odds_string <- paste0(OR, " (",odds_CI_2.5,"-",odds_CI_97.5,")")
```
Make summary table of descriptive data for continuous variables

```{r}
summary <- data.frame(cbind("name" = names(p_value),
  "patients_who_survived" = string_means$survived_mean, 
  "patients_who_died" = string_means$died_mean, 
  "overall" = string_means$all_mean,
  "missing" = missing_string, 
  "p_value" = p_value, 
  "odds_ratio" = odds_string))

summary
# print summary table with most significant p values first
# arrange(summary, p_value)
# attach(summary)
# cbind(name, odds_ratio, p_value)
# detach(summary)
dir.create("lassa_fever_analysis")
write.csv(summary, "/Users/kelsey/Dropbox (UMass Medical School)/Kelsey/Lassa Fever/Data/lassa_fever_analysis/continuous_variable_summary.csv")
```

Calculate CFR by age group.

```{r}
ag <- data.frame("age" = cont$age, "out" = cont$out)
ag <- na.omit(ag)
labs <- c("0-10", "10-20", "20-30", "30-40", "40-50", "50-60", "60-70", "70+")
ag$group <- cut(ag$age, breaks = c(seq(0, 70, by = 10), Inf), labels = labs, right = FALSE)

ag <- ag %>%
  group_by(group) %>%
  dplyr::summarize(total = n(), surv = table(out)[1], died = table(out)[2], CFR = table(out)[2]/n())
write.csv(ag, "/Users/kelsey/Dropbox (UMass Medical School)/Kelsey/Lassa Fever/Data/lassa_fever_analysis/cfr_by_age.csv")
```

Is viral load >20 associated with increased risk of death?
```{r}
cont$pcr_bin <- ifelse(cont$pcrct > 25, 0, 1)
glm(out ~ pcr_bin, data = cont, family = binomial)
cont
exp(2.342)
```
