==> 7_model-calibration.Rmd <==
TO DO:

- calculate LPs for validation
- Re-estimate intercept and slope
```{r}
setwd("/Users/kelsey/Dropbox (UMass Medical School)/Kelsey/Lassa Fever/Data")
lassa <- read.csv("2018-2021-cleaned.csv")
lassa <- lassa
```

Model 1: 
intercept -4.311156675 
age 0.035277727 
scns 1.429050048 
bleeding 0.768763947 
ast 0.003146318 
cre 0.174740591 

Model 2:
intercept -7.41300283 
age 0.03801523 
scns 1.03407785 
bleeding 0.95297837 
cre 0.16937162 
k 0.85508951 

```{r}
model1_coef <- list(-4.311156675, 0.035277727, 1.429050048, 0.768763947, 0.003146318, 0.174740591)
model2_coef <-  list(-7.41300283, 0.03801523, 1.03407785, 0.95297837, 0.16937162, 0.85508951)

one <- rep(1, nrow(lassa))
model1data <- data.frame("int" = one, 
                         "age" = lassa$age, 
                         "scns" = lassa$scns, 
                         "blding" = lassa$blding, 
                         "ast" = lassa$ast, 
                         "cre" = lassa$cre, 
                         "out" = lassa$out)

model2data <- data.frame("int" = one, 
                         "age" = lassa$age, 
                         "scns" = lassa$scns, 
                         "blding" = lassa$blding, 
                         "cre" = lassa$cre, 
                         "k" = lassa$k, 
                         "out" = lassa$out)

model1pred <- rowSums(na.omit(model1data)[-7]*model1_coef)
model2pred <- rowSums(na.omit(model2data)[-7]*model2_coef)

model1prob <- exp(model1pred)/ (1 + exp(model1pred))

str(model1pred)
na.omit(model1data)*model1_coef
```
